Split subagent result rendering out of the extension body
The renderers were the largest self-contained block in the extension entry
point and the only code there that touches the tui widgets. Moving them, plus
the result shapes they share with the tool, leaves the factory with the schema,
the dispatch and the session hooks.

Pure move: no behavior change.
